moved credentials file to config folder
