+++ test/_setup/config/.gitignore
@@ -1,0 +1,1 @@
+credentials.js
